requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
# pymupdf
# pdfplumber
google-cloud-storage>=2.16.0
google-genai>=1.66.0
google-adk>=1.26.0
